LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-order-2-import-expected.txt

CONSOLE MESSAGE: Error: error

FAIL Test evaluation order of modules assert_array_equals: expected property 4 to be "microtask-2.2" but got "script-load" (expected array ["step-2.2-1", "step-2.2-2", "global-error", "error", "microtask-2.2", "script-load", "global-load"] got ["step-2.2-1", "step-2.2-2", "global-error", "error", "script-load", "global-load", "microtask-2.2"])
